divided battery services into single capabilities
