+++ example/lib/widgets/battery_info_widget.dart
@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:open_earable_flutter/open_earable_flutter.dart';
 
 class BatteryInfoWidget extends StatelessWidget {
-  final ExtendedBatteryService connectedDevice;
+  final Wearable connectedDevice;
 
   const BatteryInfoWidget({super.key, required this.connectedDevice});
 
@@ -14,20 +14,22 @@ class BatteryInfoWidget extends StatelessWidget {
       child: Column(
         crossAxisAlignment: CrossAxisAlignment.start,
         children: [
+          if (connectedDevice is BatteryLevelService)
+            StreamBuilder(
+              stream: (connectedDevice as BatteryLevelService).batteryPercentageStream,
+              builder: (context, snapshot) {
+                if (snapshot.hasData) {
+                  return Text(
+                    "Battery Percentage:\t${snapshot.data}%",
+                  );
+                } else {
+                  return const CircularProgressIndicator();
+                }
+              },
+            ),
+          if (connectedDevice is BatteryLevelStatusService)
           StreamBuilder(
-            stream: connectedDevice.batteryPercentageStream,
-            builder: (context, snapshot) {
-              if (snapshot.hasData) {
-                return Text(
-                  "Battery Percentage:\t${snapshot.data}%",
-                );
-              } else {
-                return const CircularProgressIndicator();
-              }
-            },
-          ),
-          StreamBuilder(
-            stream: connectedDevice.powerStatusStream,
+            stream: (connectedDevice as BatteryLevelStatusService).powerStatusStream,
             builder: (context, snapshot) {
               if (snapshot.hasData) {
                 return Column(
@@ -62,53 +64,55 @@ class BatteryInfoWidget extends StatelessWidget {
             },
           ),
           const Divider(),
-          StreamBuilder(
-            stream: connectedDevice.healthStatusStream,
-            builder: (context, snapshot) {
-              if (snapshot.hasData) {
-                return Column(
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: [
-                    Text(
-                      "Health Summary: ${snapshot.data!.healthSummary}%",
-                    ),
-                    Text(
-                      "Cycle Count: ${snapshot.data!.cycleCount}",
-                    ),
-                    Text(
-                      "Current Temperature: ${snapshot.data!.currentTemperature}°C",
-                    ),
-                  ],
-                );
-              } else {
-                return const CircularProgressIndicator();
-              }
-            },
-          ),
+          if (connectedDevice is BatteryHealthStatusService)
+            StreamBuilder(
+              stream: (connectedDevice as BatteryHealthStatusService).healthStatusStream,
+              builder: (context, snapshot) {
+                if (snapshot.hasData) {
+                  return Column(
+                    crossAxisAlignment: CrossAxisAlignment.start,
+                    children: [
+                      Text(
+                        "Health Summary: ${snapshot.data!.healthSummary}%",
+                      ),
+                      Text(
+                        "Cycle Count: ${snapshot.data!.cycleCount}",
+                      ),
+                      Text(
+                        "Current Temperature: ${snapshot.data!.currentTemperature}°C",
+                      ),
+                    ],
+                  );
+                } else {
+                  return const CircularProgressIndicator();
+                }
+              },
+            ),
           const Divider(),
-          StreamBuilder(
-            stream: connectedDevice.energyStatusStream,
-            builder: (context, snapshot) {
-              if (snapshot.hasData) {
-                return Column(
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: [
-                    Text(
-                      "Voltage: ${snapshot.data!.voltage}V",
-                    ),
-                    Text(
-                      "Available Capacity: ${snapshot.data!.availableCapacity}Wh",
-                    ),
-                    Text(
-                      "Charge Rate: ${snapshot.data!.chargeRate}W",
-                    ),
-                  ],
-                );
-              } else {
-                return const CircularProgressIndicator();
-              }
-            },
-          ),
+          if (connectedDevice is BatteryEnergyStatusService)
+            StreamBuilder(
+              stream: (connectedDevice as BatteryEnergyStatusService).energyStatusStream,
+              builder: (context, snapshot) {
+                if (snapshot.hasData) {
+                  return Column(
+                    crossAxisAlignment: CrossAxisAlignment.start,
+                    children: [
+                      Text(
+                        "Voltage: ${snapshot.data!.voltage}V",
+                      ),
+                      Text(
+                        "Available Capacity: ${snapshot.data!.availableCapacity}Wh",
+                      ),
+                      Text(
+                        "Charge Rate: ${snapshot.data!.chargeRate}W",
+                      ),
+                    ],
+                  );
+                } else {
+                  return const CircularProgressIndicator();
+                }
+              },
+            ),
         ],
       ),
     );
