+++ lib/open_earable_flutter.dart
@@ -19,7 +19,10 @@ export 'src/models/devices/wearable.dart';
 export 'src/models/capabilities/device_firmware_version.dart';
 export 'src/models/capabilities/device_hardware_version.dart';
 export 'src/models/capabilities/device_identifier.dart';
-export 'src/models/capabilities/battery_service.dart';
+export 'src/models/capabilities/battery_level.dart';
+export 'src/models/capabilities/battery_level_status.dart';
+export 'src/models/capabilities/battery_health_status.dart';
+export 'src/models/capabilities/battery_energy_status.dart';
 export 'src/models/capabilities/rgb_led.dart';
 export 'src/models/capabilities/status_led.dart';
 export 'src/models/capabilities/sensor.dart';
