+++ lib/src/models/devices/open_earable_v2.dart
@@ -3,6 +3,10 @@ import 'dart:math';
 import 'dart:typed_data';
 
 import 'package:open_earable_flutter/open_earable_flutter.dart';
+import 'package:open_earable_flutter/src/models/capabilities/battery_energy_status.dart';
+import 'package:open_earable_flutter/src/models/capabilities/battery_health_status.dart';
+import 'package:open_earable_flutter/src/models/capabilities/battery_level.dart';
+import 'package:open_earable_flutter/src/models/capabilities/battery_level_status.dart';
 import '../../managers/ble_manager.dart';
 
 const String _batteryLevelCharacteristicUuid = "2A19";
@@ -29,7 +33,10 @@ class OpenEarableV2 extends Wearable
         SensorConfigurationManager,
         RgbLed,
         StatusLed,
-        ExtendedBatteryService,
+        BatteryLevelService,
+        BatteryLevelStatusService,
+        BatteryHealthStatusService,
+        BatteryEnergyStatusService,
         DeviceIdentifier,
         DeviceFirmwareVersion,
         DeviceHardwareVersion {
